In FramedGraph: Don't call getBaseGraph()
Calling getBaseGraph() and operating directly on that in annotation
handlers for FramedGraph limits overridability possibilities. It means
that the annotation handlers bypass the methods of FramedGraph. If
those are overridden to add functionality (e.g. in a MyCustomFramedGraph
extends FramedGraph), you can't use the annotation handlers anymore.
This change fixes that.