Drop direction from the FramedGraph.frameEdge methods
It is the interface that frames an edge that defines which side is the
domain, and which side the edge. This is not something you can change
runtime, so having a parameter that allows that at runtime is useless.
The old implementation will just result in runtime class cast exceptions
when somebody passes the wrong direction parameter.

This fix has the added benefit that it's now much easier to generalize
the framing for vertices and edges alike, so that it's easy to e.g.
add support for edges to @GremlinGroovy.

See also
https://groups.google.com/forum/?fromgroups#!topic/gremlin-users/PDFkieldxFo

Conflicts:
	src/main/java/com/tinkerpop/frames/FramedGraph.java