addAdjacency id fix
When using the @Adjancency annotation with an addNewThing() method you
get a new object with the object's class as its id in the graph.  I
changed the id to null so the underlying graph would set the id instead.

I also added a test that adds two vertexes using addKnowsNewPerson.
This would have failed without the above fix, because it would try two
add two Person vertexes with the same id (Person.class).